[#12] fix: 코드 수정
