--- src/main/java/com/mobility/api/domain/transporter/service/TransporterService.java
@@ -1,7 +1,0 @@
-package com.mobility.api.domain.transporter.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class TransporterService {
-}
